First skeleton files for an ASE-based SPRKKR
This version can create a simple *.inp input file
with the possibility to add custom sections or variables.

It also create a dummy *.pot file by taking informations
from the ase.Atoms object

In the misc module, 2 helper functions can be used to
define the occupancy of any site in the structure

The test.py file in the tests folder contains an example
of basic usage.